requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community 
langchain_openai
langchain_experimental
-e . 
sentence-transformers